fix(intexfactory): best-effort cross-chain notifications for Qualified and Called transitions (#2)
* feat(intexfactory): notify BNB on series Qualified and Called transitions

Add LayerZero cross-chain notifications to OriginMessenger when a series
transitions to Qualified (qualify scan) or Called (daily called scan).
Each transition quotes the messaging fee via staticcall then sends the
message with zero msg.value; OriginMessenger self-funds from its relay float.

* fix(intexfactory): make LZ notifications best-effort in lifecycle scans

Qualify and Call are automatic begin_block / Cycle lifecycle events.
Tying them to OriginMessenger availability created a chain-level liveness
dependency: a messenger outage or exhausted relay float would prevent block
production entirely.

Extract each LZ send into a private helper (notify_lz_qualified,
notify_lz_called) and call it best-effort via `let _ = ...`. The state
transition commits regardless; BNB reconciles series state from Outbe
chain if delivery is missed.

Add tests: qualify_survives_lz_messenger_failure,
call_survives_lz_messenger_failure.

* fix(desis): resolve clippy warnings in auction module

- Tighten sort_bids parameter to &mut [BidData]
- Remove superfluous .into() on string literals in Revert construction
- Fix redundant cast in test constants

+++ crates/core/desis/src/tests.rs
@@ -93,7 +93,7 @@ fn begin_clearing_stores_pending() {
     with_storage(|s| {
         runtime::start_auction(s.clone(), SERIES_ID, default_config()).unwrap();
         runtime::reveal_auction(s.clone(), SERIES_ID, true).unwrap();
-        let supply_promis = 10 * INTEX_SIZE as u128;
+        let supply_promis = 10 * INTEX_SIZE;
         let remainder = runtime::begin_clearing(s.clone(), SERIES_ID, supply_promis).unwrap();
         assert_eq!(remainder, 0); // no rounding with exact multiple
         let contract = s.contract::<DesisContract>();
@@ -108,7 +108,7 @@ fn start_auction_derives_min_bid_qty_from_prior_clearing() {
         // First auction: start, reveal, clear with 100 issued (supply = 100).
         runtime::start_auction(s.clone(), SERIES_ID, default_config()).unwrap();
         runtime::reveal_auction(s.clone(), SERIES_ID, true).unwrap();
-        runtime::begin_clearing(s.clone(), SERIES_ID, 100 * INTEX_SIZE as u128).unwrap();
+        runtime::begin_clearing(s.clone(), SERIES_ID, 100 * INTEX_SIZE).unwrap();
         runtime::process_bids_batch(
             s.clone(),
             SERIES_ID,
